+++ NEWS.md
@@ -1,5 +1,21 @@
-# cTRAP 1.6.2 (7 October, 2020)
+# cTRAP 1.8 (22 October, 2020)
 
+## Interactive functions for loading data and analysing results
+
+* New Shiny-based graphical interface functions:
+    - `launchDiffExprLoader()`: load differential expression data
+    - `launchCMapDataLoader()`: load CMap data
+    - `launchResultPlotter()`: view and plot data results
+    - `launchMetadataViewer()`: check metadata of a given object
+    
+## Major changes
+
+* `downloadENCODEknockdownMetadata()`: metadata is automatically saved to a file
+in order to avoid downloading metadata every time this function is run
+* `plotTargetingDrugsVSsimilarPerturbations()`:
+    - automatically look for matching compounds in multiple columns of both
+    datasets
+    - allow to manually select columns on which to merge datasets
 * Simplify tutorial
 
 # cTRAP 1.6.1 (17 August, 2020)